local-skill-scenario-1/requirements.txt
langgraph
langchain-openai
python-dotenv
azure-identity
pyyaml

azure-ai-agentserver-core>=1.0.0b15
azure-ai-agentserver-langgraph>=1.0.0b15
